Provide mariadb 10.1 and 10.2, fixes #1221 (#1318)

+++ containers/ddev-dbserver/test/testserver.sh
@@ -3,7 +3,7 @@ set -x
 set -euo pipefail
 
 IMAGE="$1"  # Full image name with tag
-MYSQL_VERSION="$2"
+MARIADB_VERSION="$2"
 CONTAINER_NAME="testserver"
 HOSTPORT=33000
 MYTMPDIR="${HOME}/tmp/testserver-sh_${RANDOM}_$$"
@@ -30,12 +30,14 @@ function cleanup {
 
 # Wait for container to be ready.
 function containercheck {
+    set +x
 	for i in {60..0};
 	do
 		# status contains uptime and health in parenthesis, sed to return health
 		status="$(docker ps --format "{{.Status}}" --filter "name=$CONTAINER_NAME" | sed  's/.*(\(.*\)).*/\1/')"
 		if [[ "$status" == "healthy" ]]
 		then
+		    set -x
 			return 0
 		fi
 		sleep 1
@@ -82,10 +84,10 @@ RES=$?
 if [ $RES -eq 0 ]; then
 	echo "Successful mysql show variables, output=$OUTPUT"
 fi
-versionregex="version	$MYSQL_VERSION"
+versionregex="version	$MARIADB_VERSION"
 if [[ $OUTPUT =~ $versionregex ]];
 then
-	echo "Version check ok - found '$MYSQL_VERSION'"
+	echo "Version check ok - found '$MARIADB_VERSION'"
 else
 	echo "Expected to see $versionregex. Actual output: $OUTPUT"
 	exit 105
@@ -132,30 +134,30 @@ fi
 
 cleanup
 
+if [ "$MARIADB_VERSION" != "10.1" ] ; then
+    # Start up with a Mariadb 10.1 database and verify that it gets upgraded successfully to 10.2
+    docker volume rm $VOLUME && docker volume create $VOLUME
+    # Populate the volume with the contents of our 10.1 tarball. Here it doesn't matter that
+    # we're putting it in /var/lib/mysql, but it's put there just for clarity of purpose.
+    docker run -i --rm -v "$VOLUME:/var/lib/mysql" busybox tar -C //var/lib/mysql -zxf - <test/testdata/d6git_basic_mariadb_10_1.tgz
+    # Now start up the container with the populated volume
+    if ! docker run -u "$MOUNTUID:$MOUNTGID" -v "$VOLUME:/var/lib/mysql" --name=$CONTAINER_NAME -d $IMAGE; then
+        echo "MySQL server start failed with error code $?"
+        exit 109
+    fi
+    containercheck
+    # Here we should show an upgrade happening because this was a 10.1 database
+    (docker logs $CONTAINER_NAME 2>&1 | grep "Running mysql_upgrade because my_mariadb_version=10.2 is not the same as db_mariadb_version=10.1" >/dev/null 2>&1) || (echo "Failed to find mysql_upgrade clause in docker logs" && exit 4)
 
-# Start up with a Mariadb 10.1 database and verify that it gets upgraded successfully to 10.2
-docker volume rm $VOLUME && docker volume create $VOLUME
-# Populate the volume with the contents of our 10.1 tarball. Here it doesn't matter that
-# we're putting it in /var/lib/mysql, but it's put there just for clarity of purpose.
-docker run -i --rm -v "$VOLUME:/var/lib/mysql" busybox tar -C //var/lib/mysql -zxf - <test/testdata/d6git_basic_mariadb_10_1.tgz
-# Now start up the container with the populated volume
-if ! docker run -u "$MOUNTUID:$MOUNTGID" -v "$VOLUME:/var/lib/mysql" --name=$CONTAINER_NAME -d $IMAGE; then
-	echo "MySQL server start failed with error code $?"
-	exit 109
+    docker stop $CONTAINER_NAME && docker rm $CONTAINER_NAME
+    # Now run the container again with the same volume (now upgraded) and make sure we don't have the upgrade action
+    if ! docker run -u "$MOUNTUID:$MOUNTGID" -v "$VOLUME:/var/lib/mysql" --name=$CONTAINER_NAME -d $IMAGE; then
+        echo "MySQL server start failed with error code $?"
+        exit 110
+    fi
+    containercheck
+    (docker logs $CONTAINER_NAME 2>&1 | grep -v "Running mysql_upgrade" >/dev/null 2>&1) || (echo "Found  mysql_upgrade clause in docker logs when upgrade should not have happened" && exit 6)
 fi
-containercheck
-# Here we should show an upgrade happening because this was a 10.1 database
-(docker logs $CONTAINER_NAME 2>&1 | grep "Running mysql_upgrade because my_mariadb_version=10.2 is not the same as db_mariadb_version=10.1" >/dev/null 2>&1) || (echo "Failed to find mysql_upgrade clause in docker logs" && exit 4)
-
-docker stop $CONTAINER_NAME && docker rm $CONTAINER_NAME
-# Now run the container again with the same volume (now upgraded) and make sure we don't have the upgrade action
-if ! docker run -u "$MOUNTUID:$MOUNTGID" -v "$VOLUME:/var/lib/mysql" --name=$CONTAINER_NAME -d $IMAGE; then
-	echo "MySQL server start failed with error code $?"
-	exit 110
-fi
-containercheck
-(docker logs $CONTAINER_NAME 2>&1 | grep -v "Running mysql_upgrade" >/dev/null 2>&1) || (echo "Found  mysql_upgrade clause in docker logs when upgrade should not have happened" && exit 6)
-
 
 echo "Tests passed"
 exit 0
